Deprecated/Removed index and aggregation_variables modules
